log_service/pyproject.toml
[tool.poetry]
name = "log-service"
version = "0.1.0"
description = ""
authors = ["Jake <nhnam6@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
dev = "uvicorn main:app --host 0.0.0.0 --port 8000 --reload --log-level debug"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.13"
uvicorn = {extras = ["standard"], version = "^0.34.3"}
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
pydantic = "^2.11.7"
boto3 = "^1.38.44"
opensearch-py = "^3.0.0"
pydantic-settings = "^2.10.1"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
alembic = "^1.16.2"
pytest = "^8.4.1"
httpx = "^0.28.1"
pytest-cov = "^6.2.1"
coverage = "^7.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = ".venv,__pycache__,build,dist"
